WIP: initial attempt to detect new objects
The idea is for each pixel in the selected image, search for the
most similar pixel (in terms of intensity) in the projected
reference image. The absolute difference of the found intensity and
the original intensity is some sort of indication whether the pixel
exists in the reference or not.